annotation_frontend/requirements.txt
# Annotation Frontend Service - Pinned Dependencies
# Last updated: June 2025
# Pin critical versions for reproducible builds and security fixes

# Core web framework and ASGI server - Security Updates
fastapi==0.115.6
uvicorn[standard]==0.32.1
starlette==0.41.3

# Rate limiting
slowapi==0.1.9

# Templating and form handling - Security Updates
jinja2==3.1.6
python-multipart==0.0.18

# Kafka messaging
confluent-kafka==2.10.0
aiokafka==0.11.0

# Redis caching and queue management
redis>=5.0.0
upstash-redis>=0.15.0

# Data validation and settings
pydantic==2.11.5
pydantic-settings==2.1.0

# Logging and monitoring
structlog==25.3.0
prometheus_client==0.22.0
python-json-logger==3.3.0
psutil==6.1.1

# Authentication and security - Security Updates
PyJWT==2.10.1
python-jose[cryptography]==3.4.0
fastapi-csrf-protect==0.3.4
python-multipart==0.0.18

# Testing framework
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-mock==3.14.0
testcontainers==4.8.2

# HTTP client for testing
httpx==0.28.1

# Security validation (for CI/CD)
pip-audit==2.9.0